requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
asgiref==3.3.4; python_version >= '3.6'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
beautifulsoup4==4.9.3
dj-database-url==0.5.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-hamlpy==1.4.3
django-heroku==0.3.1
django==3.2.2
djangorestframework==3.12.4
factory-boy==3.2.0
faker==8.1.3
flake8==3.9.2
gunicorn==20.1.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
psycopg2-binary==2.8.6
psycopg2==2.8.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycodestyle==2.7.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyflakes==2.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytest-django==4.2.0
pytest==6.2.4
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
regex==2020.11.13
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.2.1; python_version >= '3'
sqlparse==0.4.1; python_version >= '3.5'
text-unidecode==1.3
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
whitenoise==5.2.0
